pyproject.toml
[tool.poetry]
name = "macos-installation"
version = "0.1.0"
description = "Tool to backup / install macOS laptops"
authors = ["Scott Crooks <scott.crooks@gmail.com>"]
packages = [{include = "macos_installation", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
coloredlogs = "^15.0.1"
pycryptodomex = "^3.16.0"
tabulate = "^0.9.0"

[tool.poetry.scripts]
macos-install = "macos_installation.main:main"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
ipython = "^8.7.0"
isort = "^5.11.4"
mypy = "^0.991"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
